theme parser object added
